some where ahead

+++ level3/local-environment.hh
@@ -72,7 +72,7 @@ public:
 
 class Eval_Result_Value_Int:public Eval_Result_Value
 {
-	int value;
+	float value;
 	bool defined;
 public:
 	Eval_Result_Value_Int();
